requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
azure-functions-openapi
packaging
azure-ai-inference
openai
azure-search-documents>=11.6.0b4
python-dotenv